+++ sport-log-client/lib/pages/workout/metcon_sessions/metcon_session_edit_page.dart
@@ -65,6 +65,10 @@ class _MetconSessionEditPageState extends State<MetconSessionEditPage> {
   }
 
   Future<void> _deleteMetconSession() async {
+    final delete = await showDeleteWarningDialog(context, "Metcon Session");
+    if (!delete) {
+      return;
+    }
     if (!widget.isNew) {
       await _dataProvider.deleteSingle(_metconSessionDescription);
     }
